added data friles from EMM, along with .mtg and .tol files for testing
