+++ .gitignore
@@ -1,2 +1,4 @@
 /venv
 .ipynb_checkpoints/
+*.DS_Store
+__pycache__
