+++ data/emm0311.skp
@@ -1,0 +1,3 @@
+ 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
+ 1 0 1 0 0 1 1 1 1 1 1 1 1 1 1 1
+ 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6
