+++ data/emm0311.tol
@@ -1,0 +1,171 @@
+LA02	-3.000000	73.043478	53.444444
+LA04	-4.000000	58.840580	67.833333
+LA06	-2.000000	43.623188	78.111111
+LA09	-5.000000	19.275362	86.333333
+LA11	-7.000000	-1.586207	91.472222
+LA13	-6.000000	-16.655172	95.583333
+LA16	-4.000000	-45.282051	93.527778
+LA18	-6.000000	-65.538462	86.333333
+LA20	-7.000000	-81.743590	76.055556
+LA22	-8.000000	-92.884615	60.638889
+LA25	-8.000000	-105.038462	38.027778
+LA27	-10.000000	-109.089744	18.500000
+LA30	-11.000000	-113.141026	-11.550000
+LA32	-13.000000	-112.128205	-33.600000
+LB01	-13.000000	67.971014	56.527778
+LB04	-16.000000	19.275362	84.277778
+LB07	-21.000000	-32.115385	92.500000
+LB11	-21.000000	-96.935897	51.388889
+LB14	-26.000000	-110.102564	-4.200000
+LC02	-26.000000	47.681159	67.833333
+LC06	-28.000000	-12.689655	88.388889
+LC09	-31.000000	-68.576923	78.111111
+LC12	-33.000000	-99.974359	37.000000
+LC15	-36.000000	-106.051282	-22.050000
+LD02	-35.000000	58.840580	53.444444
+LD04	-34.000000	45.652174	63.722222
+LD06	-33.000000	32.463768	70.916667
+LD09	-32.000000	10.144928	80.166667
+LD11	-33.000000	-6.344828	84.277778
+LD13	-36.000000	-19.961538	86.333333
+LD15	-40.000000	-38.192308	85.305556
+LD17	-40.000000	-59.461538	77.083333
+LD20	-42.000000	-81.743590	59.611111
+LD23	-42.000000	-94.910256	38.027778
+LD26	-45.000000	-97.948718	14.388889
+LD28	-46.000000	-99.974359	-4.200000
+LD30	-50.000000	-99.974359	-24.150000
+LE01	-46.000000	56.811594	46.250000
+LE03	-42.000000	43.623188	58.583333
+LE06	-39.000000	25.362319	69.888889
+LE08	-40.000000	10.144928	76.055556
+LE10	-42.000000	-4.758621	79.138889
+LE13	-47.000000	-27.051282	81.194444
+LE15	-50.000000	-45.282051	78.111111
+LE17	-51.000000	-63.512821	69.888889
+LE19	-51.000000	-76.679487	57.555556
+LE21	-51.000000	-86.807692	42.138889
+LE23	-50.000000	-92.884615	26.722222
+LE25	-51.000000	-94.910256	8.222222
+LE27	-56.000000	-93.897436	-6.300000
+LE29	-57.000000	-92.884615	-27.300000
+LF02	-49.000000	33.478261	57.555556
+LF05	-53.000000	-7.137931	70.916667
+LF08	-61.000000	-54.397436	63.722222
+LF11	-60.000000	-83.769231	24.666667
+LF14	-64.000000	-80.730769	-24.150000
+LG01	-58.000000	35.507246	46.250000
+LG04	-58.000000	0.000000	63.722222
+LG07	-67.000000	-41.230769	58.583333
+LG10	-67.000000	-72.628205	24.666667
+LG13	-68.000000	-72.628205	-18.900000
+LH02	-65.000000	14.202899	48.305556
+LH04	-67.000000	-7.931034	54.472222
+LH06	-73.000000	-34.141026	48.305556
+LH08	-74.000000	-55.410256	33.916667
+LH10	-71.000000	-67.564103	11.305556
+LH12	-70.000000	-64.525641	-18.900000
+LH13	-70.000000	-59.461538	-34.650000
+LI01	-70.000000	23.333333	26.722222
+LI03	-72.000000	2.028986	39.055556
+LI05	-76.000000	-17.448276	42.138889
+LI07	-78.000000	-40.217949	32.888889
+LI09	-76.000000	-57.435897	15.416667
+LI11	-74.000000	-56.423077	-8.400000
+LI13	-77.000000	-43.256410	-35.700000
+LJ01	-75.000000	10.144928	22.611111
+LJ04	-79.000000	-39.205128	15.416667
+LK02	-79.000000	-5.551724	13.361111
+LK05	-78.000000	-35.153846	-19.950000
+LL01	-75.000000	11.159420	-17.850000
+LL02	-77.000000	6.086957	-6.300000
+LL03	-79.000000	-1.586207	1.027778
+LL04	-80.000000	-10.310345	1.027778
+LL05	-80.000000	-18.241379	-2.100000
+LL06	-79.000000	-21.987179	-10.500000
+RA01	9.134328	77.101449	45.222222
+RA02	9.134328	68.985507	56.527778
+RA04	8.119403	52.753623	70.916667
+RA07	8.119403	30.434783	82.222222
+RA09	8.119403	14.202899	86.333333
+RA11	9.134328	-3.172414	90.444444
+RA13	6.089552	-18.948718	94.555556
+RA15	3.044776	-36.166667	95.583333
+RA17	3.044776	-54.397436	90.444444
+RA20	3.044776	-81.743590	76.055556
+RA22	3.044776	-92.884615	61.666667
+RA25	1.014925	-106.051282	33.916667
+RA27	0.000000	-110.102564	17.472222
+RA29	-1.000000	-112.128205	-2.100000
+RA31	-3.000000	-113.141026	-25.200000
+RB02	17.253731	46.666667	71.944444
+RB06	17.253731	-16.655172	91.472222
+RB09	15.223881	-69.589744	84.277778
+RB12	12.179104	-106.051282	34.944444
+RB15	9.134328	-112.128205	-18.900000
+RC01	30.447761	59.855072	54.472222
+RC04	28.417910	12.173913	80.166667
+RC07	25.373134	-33.128205	89.416667
+RC11	22.328358	-94.910256	51.388889
+RC14	19.283582	-110.102564	-4.200000
+RD01	43.641791	57.826087	47.277778
+RD03	37.552239	45.652174	60.638889
+RD05	36.537313	30.434783	68.861111
+RD07	33.492537	16.231884	76.055556
+RD09	32.477612	-0.793103	81.194444
+RD11	34.507463	-13.482759	83.250000
+RD13	36.537313	-30.089744	85.305556
+RD15	36.537313	-49.333333	82.222222
+RD17	37.552239	-65.538462	77.083333
+RD20	36.537313	-84.782051	58.583333
+RD23	32.477612	-98.961538	31.861111
+RD25	31.462687	-102.000000	18.500000
+RD28	29.432836	-106.051282	-8.400000
+RD30	33.492537	-103.012821	-28.350000
+RE02	46.686567	39.565217	56.527778
+RE04	44.656716	24.347826	64.750000
+RE06	43.641791	10.144928	70.916667
+RE09	41.611940	-11.103448	78.111111
+RE11	44.656716	-24.012821	80.166667
+RE13	43.641791	-42.243590	80.166667
+RE15	44.656716	-61.487179	74.000000
+RE17	44.656716	-74.653846	64.750000
+RE20	38.567164	-89.846154	46.250000
+RE23	38.567164	-96.935897	20.555556
+RE26	38.567164	-99.974359	-4.200000
+RE28	37.552239	-100.987179	-22.050000
+RE30	38.567164	-98.961538	-38.850000
+RF01	54.805970	30.434783	52.416667
+RF04	49.731343	-8.724138	70.916667
+RF07	54.805970	-55.410256	67.833333
+RF10	47.701493	-87.820513	31.861111
+RF13	41.611940	-97.948718	-12.600000
+RG02	61.910448	8.115942	52.416667
+RG05	60.895522	-29.076923	63.722222
+RG08	61.910448	-65.538462	47.277778
+RG11	50.746269	-86.807692	10.277778
+RG14	52.776119	-82.756410	-38.850000
+RH01	68.000000	10.144928	41.111111
+RH03	66.985075	-11.896552	51.388889
+RH05	68.000000	-38.192308	50.361111
+RH07	65.970149	-60.474359	41.111111
+RH09	58.865672	-76.679487	16.444444
+RH11	58.865672	-78.705128	-8.400000
+RH13	59.880597	-71.615385	-34.650000
+RI02	72.059701	-1.586207	37.000000
+RI04	72.059701	-18.948718	45.222222
+RI06	72.059701	-42.243590	39.055556
+RI08	69.014925	-61.487179	23.638889
+RI10	63.940299	-70.602564	1.027778
+RI12	64.955224	-66.551282	-29.400000
+RI13	62.925373	-59.461538	-44.100000
+RJ02	77.134328	-21.987179	28.777778
+RJ05	68.000000	-64.525641	-7.350000
+RK01	77.134328	-12.689655	9.250000
+RK04	70.029851	-54.397436	-4.200000
+RL01	73.074627	-6.344828	-24.150000
+RL02	77.134328	-12.689655	-6.300000
+RL03	78.149254	-17.935897	0.000000
+RL04	77.134328	-29.076923	0.000000
+RL06	73.074627	-42.243590	-7.350000
+RL07	77.134328	-43.256410	-25.200000
